requirements.txt
# ========================================
# RAGLite 完整依赖包列表
# ========================================

# 核心AI框架
langchain==0.2.3
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-chroma==0.1.1
openai==1.3.0

# Web应用框架
streamlit==1.36.0

# 数据库相关
pymysql==1.1.0
sqlalchemy==2.0.31
chromadb>=0.4.15

# 文档处理
pypdf==4.2.0
python-docx==1.1.0
tiktoken==0.7.0
chardet>=5.2.0

# 环境配置
python-dotenv>=1.0.0

# 数据处理和分析
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# HTTP请求
requests>=2.31.0

# 类型检查和验证
pydantic>=2.0.0
typing-extensions>=4.8.0

# 数据可视化 (可选)
matplotlib>=3.7.0
seaborn>=0.12.0

# 机器学习工具 (可选)
scikit-learn>=1.3.0
sentence-transformers>=2.2.2

# 高性能向量检索 (可选)
faiss-cpu>=1.7.4

# 开发和测试工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# 日志工具
loguru>=0.7.0

# 安全相关
cryptography>=41.0.0

# ========================================
# 安装说明:
# 
# 1. 基础安装 (必需):
#    pip install streamlit openai langchain langchain-community 
#    pip install sqlalchemy pymysql chromadb python-dotenv
#    pip install pypdf python-docx pandas numpy
#
# 2. 完整安装 (推荐):
#    pip install -r requirements.txt
#
# 3. 最小安装 (仅核心功能):
#    pip install streamlit openai langchain sqlalchemy pymysql chromadb
# ========================================
